+++ CHANGELOG.markdown
@@ -5,7 +5,7 @@
 * New `RULES` pragmas now avoid performing work for features you aren't using in a given `Traversal`, such as the index, etc.
 * Factored out a common `reversed` definition from all the various forms of it around the library and placed it in `Control.Lens.Iso`.
 * Added `binary, `octal`, `decimal` and `hex` to `Numeric.Lens`.
-* Added `\_Void` to Control.Lens.Void
+* Added `\_Void` to `Control.Lens.Prism` and `devoid` to `Control.Lens.Lens`.
 * Removed `Gettable`. We now use `Contravariant` and `Functor` together to derive `Getter` and `Fold`. This enables these to be defined
   without incurring a dependency on the `lens` package.